fix excludes and goimports  in pre-commit
Tested by moving an import in cmd/addon-operator-manager/main.go to another group  and then running pre-commit with `pre-commit run -a` to run it on everything in the repo, thus ensuring that old stuff doesn't break.

Without the fixed `exclude` regex (this is a regex - not a glob!)  the test would have also rewritten some things in `vendor/...`.

Signed-off-by: Josh Gwosdz <jgwosdz@redhat.com>