fix permissions on folder containing kind kubeconfig when it was created by kind
Kind creates the containing folder for the kubeconfig if it doesn't exist already.
With `sudo kind ...` this results in the folder being owned by rood.
This results `make delete-kind-cluster` not being able to delete the kubeconfig because a user needs write permissions on the folder to delete files from it.

Signed-off-by: Josh Gwosdz <jgwosdz@redhat.com>