Fix update after delete errors in reconciler.
This bug fixes Status updates after object deletion due to the removal
of the finalizer. As soon as the finalizer is removed the object will be
deleted from the kube-apiserver and the status update call fails on
Precondition failed: UID in precondition error.

Note that the Addon is not waiting for all resources to be gone, which
is the default for kubernetes. If clients want to ensure all child
objects are garbage collected and removed, they can use foregroundDeletion.

https://kubernetes.io/docs/concepts/architecture/garbage-collection/#foreground-deletion

Signed-off-by: Nico Schieder <nschieder@redhat.com>