Add kubectl-package command
This is the cli side implementation of package-operator.
Currently it has the glorious feature set of outputting its own
version.

This commit also changes the format of image tags. Until now the
tag was just the short hash of the git commit it was based on.
Now the output of `git describe` is used (it is still possible
to force the version by setting the env VERSION).