Refactor internal/packages
Refactor the internal/packages package to make the loading process more
modular.
This PR is in preparation of adding PKO package validation to the CLI and
reports error reporting into the API as a side effect.

Example of new error output containing detailed error locations and YAML
document index within that document

Signed-off-by: Nico Schieder <nschieder@redhat.com>