Introduce test framework for package templates
Templating yaml, while being very flexible, can be error prone and
impact of changes is often hard to judge.

This PR introduces a testing framework for validating package
templating results to improve confidence and opens the door to handle
more
complex configuration inputs or topology dependend application layout.

Template test cases are a list of template inputs that will be executed
to check if the expected output is produced. After executing the
template, all other structural checks are also run.

Signed-off-by: Nico Schieder <nschieder@redhat.com>