Fix: Support long HyperShift HCP namespace names
The loader job-name was glued together from package name, namespace and
a "-loader" suffix. If Package name or namespace was very long, this
prevented the Package controller from creating the loader job.

This commit replaces the old loader job naming with UID based job names
which will always be constant in length. Package name and namespace are
set as labels to help debugging.

Signed-off-by: Nico Schieder <nschieder@redhat.com>