Remote Phase Manager lacks ObjectSet permissions
The remote phase manager lacked permissions to access ObjectSet APIs.
This access is required to lookup previous revisions to feed the
adoption checker that prevents us from working on objects already under
ownership of other Packages or controllers.

During fixing this bug, I've also found that the integration tests that
should have prevented this issue are not executed correctly skipping the
remote-phase manager.

With the problems in the integration test suite patched, we can be sure
that regressions are detected in the future :)

Signed-off-by: Nico Schieder <nschieder@redhat.com>