Hotfix: wrong obj namespace blocks teardown
If object specify a namespace that differs from the installation
namespace of a ObjectSet, teardown is blocked on failing to set
the controller reference.

This commit removes setting the controller reference in the teardown
flow and re-runs the preflight checker during teardown to prevent
deletion of objects in different namespaces.

Signed-off-by: Nico Schieder <nschieder@redhat.com>