Fix adoption not patching revision
Package Operator relies on a revision annotation to ensure ObjectSets
can negotiate ownership and don't fight over a resource.

This annotation MUST be updated at the same time the owner annotation
is or it might be reset by the next reconcilation.

Making package-operator auto-adopt it's own object causes this to
surface, although it might have been responsible for triggering other
edge-case issues.

Signed-off-by: Nico Schieder <nschieder@redhat.com>