Base package object order on fs order (#727)
* Base package object order on fs order

This commit changes the ordering of objects within a phase when
rendering a Package. Object order is now based on the order within the
filesystem.

Files are now traversed in lexical order.
e.g. a.yaml, b.yaml, c/test.yaml, cb.yaml, ...

Objects will retain the order that they are specified in each file
encountered.

* Use sort.Slice instead of sort.Sort